requirements.txt
psycopg2-binary==2.9.9
sympy==1.12
paho-mqtt==2.1.0
matplotlib==3.8.3
pandas==2.2.1